+++ src/react/view/excel/excel_styles.ts
@@ -314,7 +314,7 @@ export function excelJsCellToStyle(
             style.strike = true;
             hasStyle = true;
         }
-        if (font.underline && font.underline !== 'none' && font.underline !== false) {
+        if (font.underline && font.underline !== 'none') {
             style.underline = true;
             hasStyle = true;
         }